リポジトリを抽象化
