+++ functions/src/usecases/contracts/base-contracts.ts
@@ -1,0 +1,3 @@
+export type RepositoryAction<T, P extends unknown[]> = (
+  ...args: P
+) => Promise<T>;
